automation/requirements.txt
requests
google-genai
PyYAML
